Add a script for namespace management in Neo4j
The Neo4j Semantics plug-in allows to use RDF namespaces. Those need to
be created as properties on a node of type `NamespacePrefixDefinition`
before loading data into the graph. This script extracts the namespaces
from a RDF file and converts them into a Cypher query that will create
the required node with the namespace information.

(cherry picked from commit cc81e3f)