Add Neo4j implementation
This commit adds support for connecting to Neo4j via Spring Data Neo4j
(SDN). The configuration can the changed via the configuration file
`config/application.yml`. It uses the BOLT driver without authentication
as default.

To avoid leakage of OGM classes into the domain model a wrapping step is
done in which domain objects are created from the OGM classes. This is
done for experimental purposes and may not be a good strategy for a
production environment. Some better solution needs to be found.