Add profiles and make Neo4j the default
Because there exist several implementations and Spring needs to choose
one at startup, the `@Primary` annotation was used to pick the default
implementation. This is not very flexible because the source code needs
to be changed. This commits adds `@Profile` annotations to all available
implementations. Also, the `neo4j` profile is set as default as default
profile in the configuration file.

Closes #2.