Extend `Article` with "title" property
The `Article` class is extended to have a title property to show how the
title can be retrieved from the two back-ends. While the implementation
in SDN is straight-forward, the Blazegraph implementation required an
extension of the SPARQL query and guarding against an NPE.

The helper function to convert to `Article` is obsolete and was removed.