Update Cypher queries for retrieving statements
There was a bug that retrieving statements containing literals were not
fetched properly. It turned out that it only worked for resources by
accident. The OGM needs the full objects in the relationship entities
returned in order to properly construct the objects. Returning just the
relationship is not enough. Since some of the resources were already
loaded and therefore present in the session, they could be used to
construct the objects which lead to statements being returned.

Understanding this behavior also explains why it was working in the
tests all along. When the objects were created, they became part of the
session and already present. So returning just the relationship in the
query was enough, since all the data was in the session and could be
used to complement the constructed objects. This was not the case in the
production environment were not all data was in the session already.