Wrap CORS configuration into a filter bean
The former custom CORS configuration was provided as a
CorsConfigurationSource as recommended. Spring Security OAuth2 does not
seem to load it in the correct order, however, so that Pre-flight
Requests to the /oauth endpoints fail. Wrapping the configuration source
into a filter bean and setting the order so that the filter is applied
first solves the issue.

The former code was modified slightly. The configuration now uses
default options were possible and only applied custom configuration so
it is easier to see what is changed. It was also moved to the security
configuration instead of the application itself.