Fix test infrastructure for integration tests
This is a rather big change but required to unify the test infrastructure
and make it more reliable. The Neo4jTestContainersBaseTest is now shared
via Gradle TextFixtures, so it can be used by the REST API test code. It
was extended to wipe the database after each test method by using the
Neo4jClient, which needs to be injected. All current integration tests
where modified to use the base class, either directly or via the (badly
named) RestDocumentationBaseTest class.

There are still failing test, but those are all issues in the current
state of the code, not the test infrastructure.